scraper/requirements.txt
selenium
webdriver_manager
pillow
requests
